requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
assemblyai==0.26.0
bootstrap4==0.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-countries==7.6.1
django-crispy-forms==2.1
django_compressor==4.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
oauthlib==3.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python3-openid==3.2.0
pytube==15.0.0
rcssmin==1.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.2
rsa==4.9
sniffio==1.3.1
sqlparse==0.4.4
stripe==9.1.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
